Link between front and backend works

+++ src/main/java/mvc/mc/dh/port/in/StoryUseCase.java
@@ -5,5 +5,5 @@ import mvc.mc.dh.model.Story;
 
 public interface StoryUseCase {
     List<Story> getStories();
-    Story getStory(int id);
+    Story getStory(long id);
 }
